requirements.txt
bottle
pyusb
